feature(console): use and reconfigure global console
Use the global console returned from `rich.get_console` and reconfigure
it with the standard console settings instead of creating another
instance to prevent duplicate prints when recording is active
and one of the export_X methods is used.

This also allows others to reconfigure the console according to their
own custom preference through rich globally.

Signed-off-by: Sagi Kimhi <sagi.kimhi@wiliot.com>