feat(socx_plugins.converters): add @command converter
Add @command converter for converting a python object to a click
command.

The format is the same as for the @symbol converter, but instead of
returning the symbol, it will wrap the symbol with a `click.command()`
decorated function and return the wrapper function instead, which, once
called, will delegate the call to the imported symbol.

Signed-off-by: Sagi Kimhi <sagi.kim5@gmail.com>