feat!: add support for `env` and `fresh_env` in plugins
Add custom environments support for plugins.
Every plugin now holds an `env` field and a `fresh_env` field:

- The `env` field is a flat key-value dictionary of strings that denotes
  the environment variables of the subprocess in which the plugin will
  run upon invocation.

- The `fresh_env` field is a boolean field which denotes whether or not
  to copy variables from the environment of the main application process
  into the plugin's subprocess upon invocation of the plugin.

This is a powerful addition and was the main missing piece that
I believe was needed in order to truly make `socx` the powerful general
purpose tool it is meant to be.

There are still a few missing pieces to the puzzle that can provide
significant power boosts to `socx` such as complete control over plugins
and configurations from the cli via the `plugin` and `config` commands,
but at least from my point of view, the true essential game changer was
this feature and I've been waiting a very long time for the app to be
stable enough to add it in.

The future awaits... lets see what this app is actually capable of :)

Closes: #106
Signed-off-by: Sagi Kimhi <sagi.kim5@gmail.com>