🥳🎉The `git` gud release (v0.8.0) - `socx` and `socx git` just got gud! real gud! 🥳🎉 (#116)
* hotfix(socx_plugins.git.utils): fix find_repositories never excluding cwd

Signed-off-by: Sagi Kimhi <sagi.kim5@gmail.com>

* feat(git): run any arbitrary git command on multiple repos

Signed-off-by: Sagi Kimhi <sagi.kim5@gmail.com>

* chore(settings/plugins.yaml): reformat and organize definitions

Signed-off-by: Sagi Kimhi <sagi.kim5@gmail.com>

* feat(socx_plugins.git): support all git commands

- Support execution of any arbitrary git command on all repositories in
the manifest
- Add all manifest variables as CLI flags to allow complete control over
manifest initialization from the command line
- Add `socx.utils` sub-package and `socx.utils.decorators` for custom
socx decorators
- Add support for posix -- flag in `socx git` cli command so that all
options and arguments following it will be sent directly to the git
command being ran instead of being processed first by click in order to
allow a user to explicitly tell the app to ignore flags and just pass it
to git without first processing it in order to get around executing a
command passing options with a matching name to both `socx` and `git`
separately, e.g. `socx git -t 15 fetch -- -t # passes both -t 15 to
socx, and -t to git fetch without causing an error`

Closes: #83
Signed-off-by: Sagi Kimhi <sagi.kim5@gmail.com>

* bugfix(socx.cli.options): fix plugin.panel AttributeError in command_panels

Signed-off-by: Sagi Kimhi <sagi.kim5@gmail.com>

* feat!: add support for `env` and `fresh_env` in plugins

Add custom environments support for plugins.
Every plugin now holds an `env` field and a `fresh_env` field:

- The `env` field is a flat key-value dictionary of strings that denotes
  the environment variables of the subprocess in which the plugin will
  run upon invocation.

- The `fresh_env` field is a boolean field which denotes whether or not
  to copy variables from the environment of the main application process
  into the plugin's subprocess upon invocation of the plugin.

This is a powerful addition and was the main missing piece that
I believe was needed in order to truly make `socx` the powerful general
purpose tool it is meant to be.

There are still a few missing pieces to the puzzle that can provide
significant power boosts to `socx` such as complete control over plugins
and configurations from the cli via the `plugin` and `config` commands,
but at least from my point of view, the true essential game changer was
this feature and I've been waiting a very long time for the app to be
stable enough to add it in.

The future awaits... lets see what this app is actually capable of :)

Closes: #106
Signed-off-by: Sagi Kimhi <sagi.kim5@gmail.com>

* chore: bump version to 0.8.0

Signed-off-by: Sagi Kimhi <sagi.kim5@gmail.com>

* chore(socx_plugins.git): change arguments panel title to Git Arguments

Signed-off-by: Sagi Kimhi <sagi.kim5@gmail.com>

* bugfix(socx.config.converters): fix missing return keyword in CommandConverter

Signed-off-by: Sagi Kimhi <sagi.kim5@gmail.com>

* refactor(socx_plugins.git): remove unused dead code

Signed-off-by: Sagi Kimhi <sagi.kim5@gmail.com>

* bugfix(socx_plugins.git): fix git command timeout

Fix git command timeout not working by correcting the argument name in
socx_plugins.git.arguments from timeout to cmd_timeout so that it
updates the actual configuration value from the yaml file

Signed-off-by: Sagi Kimhi <sagi.kim5@gmail.com>

* fix(socx_plugins.git.callbacks): corret typing of manifest_lst_cb

Signed-off-by: Sagi Kimhi <sagi.kim5@gmail.com>

* bugfix(socx.config.converters): fix fresh_env being ignored by command converter

Signed-off-by: Sagi Kimhi <sagi.kim5@gmail.com>

---------

Signed-off-by: Sagi Kimhi <sagi.kim5@gmail.com>